refactor: replace custom components with React wrapper for Digdir designsystem
This major refactoring replaces manually written design system components with
a wrapper solution that uses the original React components directly. Key changes:

- Remove most custom component implementations (button, icon, multi-suggestion, etc.)
- Implement react-component-tree for rendering React components with proper style/class handling
- Update component usage throughout the app to use the new wrapper approach
- Begin DOM structure cleanup to remove unnecessary divs and improve layout
- Temporarily disable sources display in chat responses (needs refactoring)

Typography components remain for now, but will be migrated in a future commit.
This is part of a larger restructuring effort that will eventually move wrapper.cljs
and react.cljc to more appropriate locations.

Note: This commit doesn't include the minified designsystem build files, which
will need to be addressed in a future update.